Migrate broken source code to Swift 5
I am not sure whether the changes presented here are malicious to the
performance of this project. Currently, at the point of execution of
these affected lines, one (1) was the amount required to be
deinitialized and deallocate was called without params. Now we must pass
a 'count' argument to deinitialize and none to deallocate, as it appears
to be the most reasonable thing to do.